requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
transformers
torch
rouge-score
bert-score
